+++ db/migrate/20121028190413_add_raw_hash_to_feed_items.rb
@@ -1,0 +1,6 @@
+class AddRawHashToFeedItems < ActiveRecord::Migration
+  def change
+    add_column :feed_items, :raw_hash, :text, after: :url
+  end
+end
+