+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## chatter_bot 0.0.2 (unreleased)
 *   skip attachments wituout chatter_id
+*   save original raw_hash of feed_item
 
 ## chatter_bot 0.0.1 (2012-10-22)